Improvements in particle injection and Dipole
Particle charge densities are initialized with UpdateRHOcs to allow for a non constant charge injection in the boundaries.

H5hut-io debugging lines and MPI_Barriers have been lifted to improve performance.

The Dipole initialization has been slightly changed. The Bext is defined by a new subroutine in EMfields. The dipole is imposed starting from a given iteration (200 in this case) and is ramped up for a given number of iterations (1000 in the present case).

+++ include/Particles.h
@@ -25,7 +25,7 @@ developers: Stefano Markidis, Giovanni Lapenta
 class Particles {
 public:
   /** allocate particles */
-  virtual void allocate(int species, int initnpmax, Collective * col, VirtualTopology3D * vct, Grid * grid) = 0;
+  virtual void allocate(int species, long long initnpmax, Collective * col, VirtualTopology3D * vct, Grid * grid) = 0;
   /** interpolation Particle -> grid */
   virtual void interpP2G(Field * EMf, Grid * grid, VirtualTopology3D * vct) = 0;
